kiosk_server/requirements.txt
appnope==0.1.3
asgiref==3.8.1
asttokens==2.4.0
backcall==0.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
dj-rest-auth==6.0.0
django-allauth==0.63.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-restframework==0.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
executing==1.2.0
idna==3.7
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
jupyter_client==8.3.1
jupyter_core==5.3.1
matplotlib-inline==0.1.6
mysqlclient==2.2.4
nest-asyncio==1.5.8
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
pywatchman==1.4.1
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.2
tornado==6.3.3
traitlets==5.10.0
urllib3==2.2.1
wcwidth==0.2.6
